Update tuition fee model and controller for improved validation and access control
- Changed the data type of totalAmount in the tuition_fees model from Decimal to Int for better alignment with business logic.
- Enhanced the addTuitionFee and updateTuitionFee functions in studentController to include checks for existing fees based on semester and school year, preventing duplicates.
- Added access control checks to ensure only authorized users can add or update tuition fees.
- Improved error handling by providing more descriptive messages for various validation scenarios.

These changes enhance the integrity and security of the tuition fee management process.

+++ src/models/tuition_fee.js
@@ -10,7 +10,7 @@ const TuitionFee = sequelize.define(
       primaryKey: true,
     },
     studentId: { type: DataTypes.UUID, allowNull: false },
-    totalAmount: { type: DataTypes.DECIMAL(18, 2), allowNull: false },
+    totalAmount: { type: DataTypes.INTEGER, allowNull: false },
     semester: { type: DataTypes.STRING, allowNull: true },
     schoolYear: { type: DataTypes.STRING, allowNull: true },
     content: { type: DataTypes.STRING, allowNull: true },
